# maglev: Allocate permutations slice ahead of time