Previously, the Maglev permutations slice is allocated inside
`getPermutation()`, which means every time `maglev.GetLookupTable()` is
called, the slice will be recreated. That means it would recreated on
every Service creation or update.

This commit allocates the slice ahead of time when the Maglev subsystem
is initialized (`maglev.Init()`). This gives a rough improvement in time
of around 15%.

The slice is allocated based on a heuristic computation, involving the M
size. Nodes running with less than 8GB of RAM do not apply and Cilium
will not be preallocate the slice ahead of time. For nodes with more
than the aforementioned threshold, the formula for the heuristic is:

    (M / 100) * 100

This is derived from the maximum backends property from Maglev where
|backends| * 100 < M. This gives the following memory pressure profile
based on M (left-hand side):

    251:    0.004806594848632812 MB
    509:    0.019766311645507812 MB
    1021:   0.07953193664550783 MB
    2039:   0.3171936798095703 MB
    4093:   1.2781256866455077 MB
    8191:   5.118750076293945 MB
    16381:  20.472500686645507 MB
    32749:  81.82502754211426 MB
    65521:  327.5300171661377 MB
    131071: 1310.700000076294 MB

If the user has more backends than the preallocated size, we will adjust
the allocation to be |backends| * M. This is because we want to ensure
that the Maglev subsystem isn't thrashing trying to reallocate the slice
on each Service create or update.

Benchmark timings:

Before
```
$ go test -v ./pkg/maglev -check.v -check.b -check.btime 5s -check.bmem
=== RUN   Test
PASS: maglev_test.go:91: MaglevTestSuite.BenchmarkGetMaglevTable              50         341255883 ns/op1049633800 B/op        12 allocs/op
OK: 1 passed
--- PASS: Test (17.42s)
PASS
ok      github.com/cilium/cilium/pkg/maglev     17.475s
```

After
```
$ go test -v ./pkg/maglev -check.v -check.b -check.btime 5s -check.bmem
=== RUN   Test
PASS: maglev_test.go:91: MaglevTestSuite.BenchmarkGetMaglevTable              50         282792516 ns/op 1057899 B/op          11 allocs/op
OK: 1 passed
--- PASS: Test (17.89s)
PASS
ok      github.com/cilium/cilium/pkg/maglev     17.943s
```

Suggested-by: Martynas Pumputis <m@lambda.lt>

FlameGraphs:

![Before](https://user-images.githubusercontent.com/3885308/104666682-51558700-5689-11eb-941d-28bb2c124c0e.png)
![After](https://user-images.githubusercontent.com/3885308/104666723-6a5e3800-5689-11eb-9edb-fa3c5d2c4904.png)

Related: https://github.com/cilium/cilium/issues/14397