@brb 2021-01-27
> Can we derive this heuristic from elsewhere without an additional agent flag?

@borkmann I think with a large M it might be difficult to precisely predict the permutation slice size in order not to waste mem. E.g. with M=65521 the backend count (N) can be [328;644] (if we consider the 100*N < M), and thus the perm slice size is [163; 321] MB. This might be problematic for users who have backends count in the lower range.