@christarazi 2021-01-28
@borkmann @brb The agent flag has been removed and the heuristic computation has been implemented. The benchmarks numbers have stayed the same, still an improvement of 15%.